Also compute rating for higher-level agencies by averaging the ratings of their children. Also add some explanation about the rating in the page.

+++ src/main/kotlin/fr/marcsworld/model/dto/AgencyNode.kt
@@ -15,7 +15,7 @@ class AgencyNode (
 
         val active: Boolean,
 
-        val rating: Int? = null,
+        val rating: Double? = null,
 
         val documentNodes: List<DocumentNode>,
